Move existing code to a subdirectory
I believe the existing algorithms will remain useful, because there is
no way to parameterize on a type with a module, and then instantiate
that module polymorphically to have a general purpose implementation.
This can be a valid use case. For instance, there is little benefit in
specializing to the element type of a boxed vector, and timsort is
actually designed with such cases in mind (where e.g. the comparison is
not cheap). So, it seems like a backpack-based optimization strategy
should live along side the classic API, although the latter could become
less focused on super-inlining in this scenario as well.